+++ src/Cell.h
@@ -12,7 +12,7 @@ public:
     void addImage( const cv::Mat& src );
 
     //! Adds a pixel to the cell.
-    void addPixel( double gradientAngle, double weight );
+    void addPixel( double gradientAngle, float weight );
 
     //! Get the number of votes in the specified bin.
     double bin( int binIndex ) const;
@@ -21,16 +21,16 @@ public:
     double binAngle( int binIndex ) const;
 
     //! Gets the normalized [ 0, 1 ] number of votes in the specified bin.
-    double binNormalized( int binIndex ) const;
+    float binNormalized( int binIndex ) const;
 
     //! Creates a visual representation of the cell's HOG.
     cv::Mat drawHOG( int cellScale = 10 ) const;
 
     //! Gets the Cell's HOG.
-    const std::vector< double >& getHog( void ) const;
+    const std::vector< float >& getHog( void ) const;
 
     //! Gets the sum of the cell's votes.
-    double getSum( void ) const;
+    float getSum( void ) const;
 
     //! Returns the number of bins in the cell.
     int numBins( void ) const;
@@ -43,9 +43,9 @@ protected:
     static double binIndexToAngle( int binIndex, int numBins, bool shouldIgnoreSign );
 
 private:
-    std::vector< double > mBins;
+    std::vector< float > mBins;
 
-    double mSampleSum;
+    float mSampleSum;
 
     bool mShouldIgnoreSign;
